Complete Flex Exercise 2
style-copy.css is the first attempt at solving the exercise before
looking at the solution. style.css is my attempt after taking a nap
and reattempting. Solution had to be referenced again to see how to get
gap between logo and header, and to get gaps between the ul items.